The form data that we process (and replace) via the installation form page should (almost) be the same regexes that we perform replacements on when searching for potential "config.php" files for hardcoded mysql DB info variables.
Also, there are cases where the installation has to be handled in a
series of requests.  Where the user is expected to click on a link
(to continue to install.php?step=2) for example.  We now have support
for that, and we may also need to in the future add support for the
case where the GET response contains a new form (with a submit button).